backend/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
pillow
python-multipart
sqlalchemy[asyncio]
asyncmy
pydantic-settings
pydantic[email]
python-jose[cryptography]
bcrypt
aiofiles
